requirements.txt
altgraph==0.17.5
blinker==1.9.0
bottle==0.13.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clr_loader==0.2.8
colorama==0.4.6
construct==2.8.8
cryptography==46.0.3
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
oracledb==3.4.1
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
plyer==2.1.0
protobuf==6.33.2
proxy_tools==0.1.0
pycparser==2.23
pycryptodome==3.23.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.10
pymp4==1.4.0
pypiwin32==223
pystray==0.19.5
pythonnet==3.0.5
pywebview==6.1
pywidevine==1.9.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.6.2
Werkzeug==3.1.3
win10toast==0.9
